improved readability with with a bit of factorisation of 'make_pdf.js' script
